// app/components/player/jessibuca.d.ts
declare namespace Jessibuca {
  /** 超时信息 */
  enum TIMEOUT {
    /** 当play()的时候，如果没有数据返回 */
    loadingTimeout = "loadingTimeout",
    /** 当播放过程中，如果超过timeout之后没有数据渲染 */
    delayTimeout = "delayTimeout",
  }

  /** 错误信息 */
  enum ERROR {
    /** 播放错误，url 为空的时候，调用 play 方法 */
    playError = "playError",
    /** http 请求失败 */
    fetchError = "fetchError",
    /** websocket 请求失败 */
    websocketError = "websocketError",
    /** webcodecs 解码 h265 失败 */
    webcodecsH265NotSupport = "webcodecsH265NotSupport",
    /** mediaSource 解码 h265 失败 */
    mediaSourceH265NotSupport = "mediaSourceH265NotSupport",
    /** wasm 解码失败 */
    wasmDecodeError = "wasmDecodeError",
  }

  interface Config {
    /**
     * 播放器容器
     * *  若为 string ，则底层调用的是 document.getElementById('id')
     * */
    container: HTMLElement | string;
    /**
     * 设置最大缓冲时长，单位秒，播放器会自动消除延迟
     */
    videoBuffer?: number;
    /**
     * worker地址
     * *  默认引用的是根目录下面的decoder.js文件 ，decoder.js 与 decoder.wasm文件必须是放在同一个目录下面。 */
    decoder?: string;
    /**
     * 是否不使用离屏模式（提升渲染能力）
     */
    forceNoOffscreen?: boolean;
    /**
     * 是否开启当页面的'visibilityState'变为'hidden'的时候，自动暂停播放。
     */
    hiddenAutoPause?: boolean;
    /**
     * 是否有音频，如果设置`false`，则不对音频数据解码，提升性能。
     */
    hasAudio?: boolean;
    /**
     * 设置旋转角度，只支持，0(默认)，180，270 三个值
     */
    rotate?: boolean;
    /**
     * 1. 当为`true`的时候：视频画面做等比缩放后,高或宽对齐canvas区域,画面不被拉伸,但有黑边。 等同于 `setScaleMode(1)`
     * 2. 当为`false`的时候：视频画面完全填充canvas区域,画面会被拉伸。等同于 `setScaleMode(0)`
     */
    isResize?: boolean;
    /**
     * 1. 当为`true`的时候：视频画面做等比缩放后,完全填充canvas区域,画面不被拉伸,没有黑边,但画面显示不全。等同于 `setScaleMode(2)`
     */
    isFullResize?: boolean;
    /**
     * 1. 当为`true`的时候：ws协议不检验是否以.flv为依据，进行协议解析。
     */
    isFlv?: boolean;
    /**
     * 是否开启控制台调试打
     */
    debug?: boolean;
    /**
     * 1. 设置超时时长, 单位秒
     * 2. 在连接成功之前(loading)和播放中途(heart),如果超过设定时长无数据返回,则回调timeout事件
     */
    timeout?: number;
    /**
     * 1. 设置超时时长, 单位秒
     * 2. 在连接成功之前,如果超过设定时长无数据返回,则回调timeout事件
     */
    heartTimeout?: number;
    /**
     * 1. 设置超时时长, 单位秒
     * 2. 在连接成功之前,如果超过设定时长无数据返回,则回调timeout事件
     */
    loadingTimeout?: number;
    /**
     * 是否支持屏幕的双击事件，触发全屏，取消全屏事件
     */
    supportDblclickFullscreen?: boolean;
    /**
     * 是否显示网
     */
    showBandwidth?: boolean;
    /**
     * 配置操作按钮
     */
    operateBtns?: {
      /** 是否显示全屏按钮 */
      fullscreen?: boolean;
      /** 是否显示截图按钮 */
      screenshot?: boolean;
      /** 是否显示播放暂停按钮 */
      play?: boolean;
      /** 是否显示声音按钮 */
      audio?: boolean;
      /** 是否显示录制按 */
      record?: boolean;
    };
    /**
     * 开启屏幕常亮，在手机浏览器上, canvas标签渲染视频并不会像video标签那样保持屏幕常亮
     */
    keepScreenOn?: boolean;
    /**
     * 是否开启声音，默认是关闭声音播放的
     */
    isNotMute?: boolean;
    /**
     * 加载过程中文案
     */
    loadingText?: string;
    /**
     * 背景图片
     */
    background?: string;
    /**
     * 是否开启MediaSource硬解码
     * * 视频编码只支持H.264视频（Safari on iOS不支持）
     * * 不支持 forceNoOffscreen 为 false (开启离屏渲染)
     */
    useMSE?: boolean;
    /**
     * 是否开启Webcodecs硬解码
     * *  视频编码只支持H.264视频 (需在chrome 94版本以上，需要https或者localhost环境)
     * *  支持 forceNoOffscreen 为 false （开启离屏渲染)
     * */
    useWCS?: boolean;
    /**
     * 是否开启键盘快捷键
     * 目前支持的键盘快捷键有：esc -> 退出全屏；arrowUp -> 声音增加；arrowDown -> 声音减少；
     */
    hotKey?: boolean;
    /**
     *  在使用MSE或者Webcodecs 播放H265的时候，是否自动降级到wasm模式。
     *  设置为false 则直接关闭播放，抛出Error 异常，设置为true 则会自动切换成wasm模式播放。
     */
    autoWasm?: boolean;
    /**
     * heartTimeout 心跳超时之后自动再播放,不再抛出异常，而直接重新播放视频地址。
     */
    heartTimeoutReplay?: boolean;
    /**
     * heartTimeoutReplay 从试次数，超过之后，不再自动播放
     */
    heartTimeoutReplayTimes?: number;
    /**
     * loadingTimeout loading之后自动再播放,不再抛出异常，而直接重新播放视频地址。
     */
    loadingTimeoutReplay?: boolean;
    /**
     * heartTimeoutReplay 从试次数，超过之后，不再自动播放
     */
    loadingTimeoutReplayTimes?: number;
    /**
     * wasm解码报错之后，不再抛出异常，而是直接重新播放视频地址。
     */
    wasmDecodeErrorReplay?: boolean;
    /**
     * https://github.com/langhuihui/jessibuca/issues/152 解决方案
     * 例如：WebGL图像预处理默认每次取4字节的数据，但是540x960分辨率下的U、V分量宽度是540/2=270不能被4整除，导致绿屏。
     */
    openWebglAlignment?: boolean;
  }
}